+++ src/main/java/co/enear/maven/plugins/keepachangelog/ValidateMojo.java
@@ -75,6 +75,8 @@ public class ValidateMojo extends InitMojo {
     public void execute() throws MojoExecutionException, MojoFailureException {
         super.execute();
         if (skip) return;
+        if (skipModules && !project.isExecutionRoot()) return;
+        if (skipRoot && project.isExecutionRoot()) return;
         Path path = getChangelogPath();
         validate(path);
     }